# docs/swagger.yaml
basePath: /
definitions:
...
  editorapi.contentAddParams:
    properties:
      copy_from_uuid:
        type: string
      copy_same_lang:
        type: boolean
      cover:
        items:
          type: string
        type: array
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      has_changed:
        type: boolean
      has_del:
        type: boolean
      is_show:
        type: boolean
      module:
        type: string
      name:
        type: string
      parent_uuid:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        type: object
      update_time:
        type: integer
      uuid:
        type: string
      version:
        type: string
    type: object
...
  editorapi.contentDelParams:
    properties:
      uuid:
        type: string
    type: object
...
  editorapi.contentEditParams:
    properties:
      content_info:
        $ref: '#/definitions/editorapi.contentInfo'
        type: object
      uuid:
        type: string
    type: object
...
  editorapi.contentInfo:
    properties:
      cover:
        items:
          type: string
        type: array
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      has_changed:
        type: boolean
      is_show:
        type: boolean
      name:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        type: object
      update_time:
        type: integer
      version:
        type: string
    type: object
  editorapi.contentListParams:
    properties:
      pageNo:
        description: 注释
        type: integer
      pageSize:
        type: integer
      parent_uuid:
        type: string
    type: object
...
  requests.ContentReportsCreateRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  requests.ContentReportsDeleteRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  requests.ContentReportsFindOneRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  requests.ContentReportsFindRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  requests.ContentReportsListRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      page_index:
        description: 分页第几页
        type: integer
      page_size:
        description: 分页数据条数
        type: integer
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      sort_type:
        description: 排序类型 1、倒叙；-1、顺序
        type: integer
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      text_field:
        description: 排序字段
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
  requests.ContentReportsUpdateRequests:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
...
  responses.ContentReportsResponses:
    properties:
      agent:
        description: agent
        type: string
      code:
        description: 课程编码
        type: string
      createdTime:
        description: 创建时间
        type: string
      dataArea:
        type: string
      dataVersion:
        description: 版本
        type: integer
      desc:
        description: 描述
        type: string
      id:
        description: ID
        type: string
      img:
        description: 图片地址
        type: string
      langCode:
        description: 归属语言
        type: string
      parentUuids:
        description: 所有父节点
        items:
          type: string
        type: array
      status:
        description: 状态 1，已处理；0，未处理；默认0
        type: integer
      tags:
        description: tags
        type: string
      userId:
        description: 用户ID
        type: string
      uuid:
        description: uuid
        type: string
    type: object
...
info:
  contact: {}
  description: 全球说编辑器后台API服务文档
  license: {}
  title: 全球说编辑器后台API服务文档
  version: 0.0.1
paths:
...
  /editor/content/version/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程内容添加,module是内容类型，现在有两个值：basic（基础内容）和levelGrade（等级测试内容）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentAddParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 课程内容版本添加
      tags:
      - EditorContentVersionAPI（内容版本接口）
  /editor/content/version/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程内容删除
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentDelParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 课程内容版本删除
      tags:
      - EditorContentVersionAPI（内容版本接口）
  /editor/content/version/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程内容修改
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentEditParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 课程内容版本修改
      tags:
      - EditorContentVersionAPI（内容版本接口）
  /editor/content/version/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 课程内容列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.contentListParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 课程内容版本列表
      tags:
      - EditorContentVersionAPI（内容版本接口）
  /editor/content_reports/create:
    post:
      consumes:
      - application/json
      description: 添加课程内容反馈
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsCreateRequests'
      produces:
      - application/jsonf
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 添加课程内容反馈
      tags:
      - 课程内容反馈接口
  /editor/content_reports/delete:
    post:
      consumes:
      - application/json
      description: 删除课程内容反馈数据
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsDeleteRequests'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除课程内容反馈数据
      tags:
      - 课程内容反馈接口
  /editor/content_reports/find:
    post:
      consumes:
      - application/json
      description: 查询课程内容反馈数据
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsFindRequests'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询课程内容反馈数据
      tags:
      - 课程内容反馈接口
  /editor/content_reports/findone:
    post:
      consumes:
      - application/json
      description: 查询一条课程内容反馈数据
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsFindOneRequests'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ContentReportsResponses'
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询一条课程内容反馈数据
      tags:
      - 课程内容反馈接口
  /editor/content_reports/list:
    post:
      consumes:
      - application/json
      description: 查询课程内容反馈列表分页数据
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsListRequests'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ContentReportsResponses'
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 查询课程内容反馈列表分页数据
      tags:
      - 课程内容反馈接口
  /editor/content_reports/update:
    post:
      consumes:
      - application/json
      description: 更新课程内容反馈数据
      parameters:
      - description: 列表参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.ContentReportsUpdateRequests'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
        "400":
          description: '{"code":500,"data":{},"msg":"失败"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新课程内容反馈数据
      tags:
      - 课程内容反馈接口
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"